Single parameter command line
This changes the CLI interaction to be a single parameter (environment
variable), the rest is assumed based on what we know.

To capture all screenshots for a given font face, now you may pass the
font name as an env variable like so:
`FONT='Fira Code' npm run screenshots`

Additionally, added `fs` as a dependency- this will allow us to check
for an existing directory, then create it if necessary.