FEAT: Simple keyboard navigation.
While playing with the site, what I really wanted was to be able
to navigate through the list without having to click on each view
individually.  This handler adds controls so that, in a desktop
environment, hovering over the navigation block activates keyboard
navigation, and 'n' and 'p' take you to the 'next' and 'prev'
fonts in the list.